requirements.txt
pyttsx3~=2.90
requests~=2.32.3
python-dotenv~=1.0.1
SpeechRecognition~=3.10.4
PyAudio~=0.2.14
google-auth~=2.32.0
google-auth-oauthlib~=1.2.1
google-auth-httplib2~=0.2.0
google-api-python-client~=2.137.0